Renamed direct value implementation to Direct*
