+++ community/values/src/main/java/org/neo4j/values/DirectFloat.java
@@ -1,0 +1,68 @@
+/*
+ * Copyright (c) 2002-2017 "Neo Technology,"
+ * Network Engine for Objects in Lund AB [http://neotechnology.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Neo4j is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package org.neo4j.values;
+
+import static java.lang.String.format;
+
+final class DirectFloat extends DirectFloatingPointNumber
+{
+    private final float value;
+
+    DirectFloat( float value )
+    {
+        this.value = value;
+    }
+
+    @Override
+    public double doubleValue()
+    {
+        return value;
+    }
+
+    @Override
+    boolean equals( boolean x )
+    {
+        return false;
+    }
+
+    @Override
+    boolean equals( char x )
+    {
+        return false;
+    }
+
+    @Override
+    boolean equals( String x )
+    {
+        return false;
+    }
+
+    @Override
+    void writeTo( ValueWriter writer )
+    {
+        writer.writeFloatingPoint( value );
+    }
+
+    @Override
+    public String toString()
+    {
+        return format( "Float(%e)", value );
+    }
+}
